pyproject.toml
[project]
name = "grnded-agents-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.75.0",
    "cryptography>=46.0.3",
    "descope>=1.7.13",
    "fastapi>=0.124.0",
    "httpx>=0.28.1",
    "mcp>=1.23.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
